+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes
 TypeTree 2.8-git
 ----------------
 
+-   The `forEachNode()` traversal function with multiple callbacks is deprecated. Use the
+    more general `applyToTree()` instead.
 -   Removed the type aliases for the storage type `Storage` and `ConstStorage` in the
     node implementations.
 -   The memberfunction `childStorage()` in the nodes now consistently take an index or an